+++ app/views/admin/admin_change_item.blade.php
@@ -1,4 +1,4 @@
-@extends('partials/layout')
+@extends('partials/admin_layout')
 @extends('partials/admin_header')
 @extends('partials/admin_sidebar')
 @extends('partials/admin_footer')
@@ -9,7 +9,7 @@
 @section('body')
 	<h1 class="admin_uni_heading">Добавить товар</h1>
 	<div class="admin_main_content">
-		{{ Form::model([], ['url'=>['/admin/update_item'], 'files'=>true, 'method'=>'POST', 'class'=>'']) }}
+		{{ Form::model($item, ['url'=>['/admin/update_item'], 'method'=>'POST', 'class'=>'']) }}
 			<div class="change_item_title_block">
 				{{ Form::label('title', 'Название: ', ['class'=>'admin_uni_lable']) }}
 				{{ Form::text('title', null, ['class'=>'', 'required']) }}
@@ -32,15 +32,15 @@
 			</div>
 			<div class="make_spec_block">
 				{{ Form::label('special', 'Добавить в спецпредложения: ', ['class'=>'admin_uni_lable']) }}
-				{{ Form::checkbox('special', false, false, ['class'=>'', 'required']) }}
+				{{ Form::checkbox('special', false, false, ['class'=>'']) }}
 			</div>
 			<div class="make_hit_block">
 				{{ Form::label('hit', 'Сделать хитом продаж: ', ['class'=>'admin_uni_lable']) }}
-				{{ Form::checkbox('hit', false, false, ['class'=>'', 'required']) }}
+				{{ Form::checkbox('hit', false, false, ['class'=>'']) }}
 			</div>
 			<div class="img_preview">
-				@if (false)
-					{{ HTML::image("img/photos/temp", "", ['class'=>'items_item_img']) }} 
+				@if (Session::get('temp'))
+					<img src='{{ URL::to("img/photos/")}}/{{ Session::get("temp") }}' class='items_item_img'>
 					<i class="fa fa-times delete_img_icon"></i>
 				@else
 					{{ HTML::image("img/photos/no_photo.png", "", ['class'=>'items_item_img']) }}
@@ -51,15 +51,19 @@
 			</div>
 			{{ Form::submit('Сохранить', ['class'=>'admin_uni_button']) }}
 		{{ Form::close() }}
+
 		<div class="change_item_img">
 			<p class="admin_uni_lable">Добавить изображение 110*95 пикс.</p>
-			{{ Form::open(['url'=>'/item_upload_image', 'method'=>'POST', 'class'=>'admin_panel_import']) }}
+			{{ Form::open(['url'=>'/admin/item_upload_image', 'files'=>true, 'method'=>'POST', 'class'=>'admin_panel_import']) }}
 				{{ Form::file('photo', ['class'=>'admin_uni_button']) }}
 				{{ Form::submit('Добавить', ['class'=>'btn admin_uni_button']) }}
 			{{ Form::close() }}	
 		</div>
-		{{ Form::open(['url'=>'/delete_item', 'method'=>'POST', 'class'=>'admin_panel_import']) }}
-			{{ Form::submit('Удалить', ['class'=>'btn admin_uni_button']) }}
-		{{ Form::close() }}	
+
+		@if ($item)
+			{{ Form::open(['url'=>"/admin/delete_item?$item->item_id", 'method'=>'POST', 'class'=>'admin_panel_import']) }}
+				{{ Form::submit('Удалить', ['class'=>'btn admin_uni_button']) }}
+			{{ Form::close() }}	
+		@endif
 	</div>
 @stop
